refactor: move email to lib/zero/
Part of #7697 (restructure src/lib/ into layered architecture).
Move lib/email/ to lib/zero/email/ since this module handles email
integration and already imports from zero-run-service.

Closes #7737
