+++ turbo/apps/web/app/api/email/unsubscribe/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import { initServices } from "../../../../src/lib/init-services";
-import { verifyUnsubscribeToken } from "../../../../src/lib/email/handlers/shared";
-import { unsubscribeUser } from "../../../../src/lib/email/unsubscribe-service";
+import { verifyUnsubscribeToken } from "../../../../src/lib/zero/email/handlers/shared";
+import { unsubscribeUser } from "../../../../src/lib/zero/email/unsubscribe-service";
 import { getAppUrl } from "../../../../src/lib/url";
 
 /**
