+++ turbo/apps/web/src/lib/zero/email/handlers/shared.ts
@@ -1,10 +1,10 @@
 import crypto from "crypto";
 import { eq } from "drizzle-orm";
-import { emailThreadSessions } from "../../../db/schema/email-thread-session";
-import { resolveDefaultAgentId } from "../../zero/resolve-default-agent";
-import { env } from "../../../env";
-import { getAppUrl } from "../../url";
-import { getApiUrl } from "../../infra/callback/dispatcher";
+import { emailThreadSessions } from "../../../../db/schema/email-thread-session";
+import { resolveDefaultAgentId } from "../../resolve-default-agent";
+import { env } from "../../../../env";
+import { getAppUrl } from "../../../url";
+import { getApiUrl } from "../../../infra/callback/dispatcher";
 import { enqueueEmail } from "../outbox-service";
 
 // ============================================================================
